+++ gpiozero/mixins.py
@@ -70,10 +70,13 @@ class SourceMixin(object):
 
     def close(self):
         try:
+            self.source = None
+        except AttributeError:
+            pass
+        try:
             super(SourceMixin, self).close()
         except AttributeError:
             pass
-        self.source = None
 
     def _copy_values(self, source):
         for v in source:
@@ -127,7 +130,7 @@ class SharedMixin(object):
     When :meth:`close` is called, an internal reference counter will be
     decremented and the instance will only close when it reaches zero.
     """
-    _INSTANCES = {}
+    _instances = {}
 
     def __del__(self):
         self._refs = 0
@@ -438,23 +441,28 @@ class HoldThread(GPIOThread):
     device is active.
     """
     def __init__(self, parent):
-        super(HoldThread, self).__init__(target=self.held, args=(parent,))
+        super(HoldThread, self).__init__(
+            target=self.held, args=(weakref.proxy(parent),))
         self.holding = Event()
         self.start()
 
     def held(self, parent):
-        while not self.stopping.is_set():
-            if self.holding.wait(0.1):
-                self.holding.clear()
-                while not (
-                        self.stopping.is_set() or
-                        parent._inactive_event.wait(parent.hold_time)
-                        ):
-                    if parent._held_from is None:
-                        parent._held_from = time()
-                    parent._fire_held()
-                    if not parent.hold_repeat:
-                        break
+        try:
+            while not self.stopping.is_set():
+                if self.holding.wait(0.1):
+                    self.holding.clear()
+                    while not (
+                            self.stopping.is_set() or
+                            parent._inactive_event.wait(parent.hold_time)
+                            ):
+                        if parent._held_from is None:
+                            parent._held_from = time()
+                        parent._fire_held()
+                        if not parent.hold_repeat:
+                            break
+        except ReferenceError:
+            # Parent is dead; time to die!
+            pass
 
 
 class GPIOQueue(GPIOThread):