+++ tests/conftest.py
@@ -1,0 +1,10 @@
+from __future__ import (
+    unicode_literals,
+    print_function,
+    absolute_import,
+    division,
+    )
+str = type('')
+
+import os
+os.environ['GPIOZERO_PIN_FACTORY'] = 'mock'